Export storage-type
