+++ src/EnumGenerator.Core/Exporter/CSharpExporter.cs
@@ -1,6 +1,5 @@
 using System;
 using System.Linq;
-using System.Reflection;
 using System.Text.RegularExpressions;
 
 using EnumGenerator.Core.Definition;
@@ -50,9 +49,9 @@ namespace EnumGenerator.Core.Exporter
             builder.WriteEmptyLine();
 
             if (string.IsNullOrEmpty(@namespace))
-                builder.AddEnum(enumDefinition);
+                builder.AddEnum(enumDefinition, storageType);
             else
-                builder.AddNamespace(@namespace, b => b.AddEnum(enumDefinition));
+                builder.AddNamespace(@namespace, b => b.AddEnum(enumDefinition, storageType));
 
             return builder.Build();
         }
@@ -73,13 +72,16 @@ namespace EnumGenerator.Core.Exporter
             builder.EndScope();
         }
 
-        private static void AddEnum(this CodeBuilder builder, EnumDefinition enumDefinition)
+        private static void AddEnum(this CodeBuilder builder, EnumDefinition enumDefinition, StorageType storageType)
         {
             var assemblyName = typeof(CSharpExporter).Assembly.GetName();
             if (!string.IsNullOrEmpty(enumDefinition.Comment))
                 builder.AddSummary(enumDefinition.Comment);
             builder.WriteLine($"[GeneratedCode(\"{assemblyName.Name}\", \"{assemblyName.Version}\")]");
-            builder.WriteLine($"public enum {enumDefinition.Identifier}");
+            if (storageType == StorageType.Implicit)
+                builder.WriteLine($"public enum {enumDefinition.Identifier}");
+            else
+                builder.WriteLine($"public enum {enumDefinition.Identifier} : {storageType.GetKeyword()}");
             builder.StartScope();
 
             var first = true;
